Do use RPATH for external dependencies.
Doing so will populate the entry `DT_RPATH` which cannot be overriden by
the users, that means that the library cannot be moved from system to
system.

When encountering an issue with external shared libraries which
libginkgo.so depends on, the recommended approach is to have a correct
`LD_LIBRARY_PATH` environment variable instead.